third_party/WebKit/LayoutTests/intersection-observer/same-document-root-expected.txt

Simple intersection observer test with explicit root and target in the same document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS entries.length is 0
PASS entries.length is 0
PASS entries.length is 1
PASS entries[0].boundingClientRect.left is 8
PASS entries[0].boundingClientRect.right is 108
PASS entries[0].boundingClientRect.top is 258
PASS entries[0].boundingClientRect.bottom is 358
PASS entries[0].intersectionRect.left is 8
PASS entries[0].intersectionRect.right is 108
PASS entries[0].intersectionRect.top is 258
PASS entries[0].intersectionRect.bottom is 308
PASS entries[0].rootBounds.left is 8
PASS entries[0].rootBounds.right is 108
PASS entries[0].rootBounds.top is 108
PASS entries[0].rootBounds.bottom is 308
PASS entries[0].target is [object HTMLDivElement]
PASS entries.length is 1
PASS entries.length is 2
PASS entries[1].boundingClientRect.left is 8
PASS entries[1].boundingClientRect.right is 108
PASS entries[1].boundingClientRect.top is 1008
PASS entries[1].boundingClientRect.bottom is 1108
PASS entries[1].intersectionRect.left is 0
PASS entries[1].intersectionRect.right is 0
PASS entries[1].intersectionRect.top is 0
PASS entries[1].intersectionRect.bottom is 0
PASS entries[1].rootBounds.left is 8
PASS entries[1].rootBounds.right is 108
PASS entries[1].rootBounds.top is 708
PASS entries[1].rootBounds.bottom is 908
PASS entries[1].target is [object HTMLDivElement]
PASS entries.length is 3
PASS entries[2].boundingClientRect.left is 8
PASS entries[2].boundingClientRect.right is 108
PASS entries[2].boundingClientRect.top is 858
PASS entries[2].boundingClientRect.bottom is 958
PASS entries[2].intersectionRect.left is 8
PASS entries[2].intersectionRect.right is 108
PASS entries[2].intersectionRect.top is 858
PASS entries[2].intersectionRect.bottom is 908
PASS entries[2].rootBounds.left is 8
PASS entries[2].rootBounds.right is 108
PASS entries[2].rootBounds.top is 708
PASS entries[2].rootBounds.bottom is 908
PASS entries[2].target is [object HTMLDivElement]
PASS successfullyParsed is true

TEST COMPLETE
